added custom timing for AI shots and changed all deltaTimes to doubles

+++ src/snorri/entities/Collider.java
@@ -23,7 +23,7 @@ public class Collider extends Entity {
 	}
 	
 	@Override
-	public void update(World world, float deltaTime) {
+	public void update(World world, double deltaTime) {
 		age += deltaTime;
 		if (age > getLifeSpan()) {
 			world.delete(this);
